requirements.txt
Django==5.0.7
djangorestframework==3.15.2
drf-yasg==1.21.7
django-filter==24.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
